# ddwrapper: Using strings.Join over string concatenation as it yields better cpu performance