```
benchmark                  old ns/op     new ns/op     delta
BenchmarkTags_Pair2-12     68.4          63.0          -7.89%
BenchmarkTags_Pair2-12     67.2          62.2          -7.44%
BenchmarkTags_Pair2-12     67.4          64.1          -4.90%
BenchmarkTags_Pair4-12     131           124           -5.34%
BenchmarkTags_Pair4-12     130           131           +0.77%
BenchmarkTags_Pair4-12     132           128           -3.03%
BenchmarkTags_Pair6-12     180           172           -4.44%
BenchmarkTags_Pair6-12     182           173           -4.95%
BenchmarkTags_Pair6-12     182           171           -6.04%
BenchmarkTags_Pair8-12     227           209           -7.93%
BenchmarkTags_Pair8-12     224           212           -5.36%
BenchmarkTags_Pair8-12     225           214           -4.89%

benchmark                  old allocs     new allocs     delta
BenchmarkTags_Pair2-12     2              2              +0.00%
BenchmarkTags_Pair2-12     2              2              +0.00%
BenchmarkTags_Pair2-12     2              2              +0.00%
BenchmarkTags_Pair4-12     3              3              +0.00%
BenchmarkTags_Pair4-12     3              3              +0.00%
BenchmarkTags_Pair4-12     3              3              +0.00%
BenchmarkTags_Pair6-12     4              4              +0.00%
BenchmarkTags_Pair6-12     4              4              +0.00%
BenchmarkTags_Pair6-12     4              4              +0.00%
BenchmarkTags_Pair8-12     5              5              +0.00%
BenchmarkTags_Pair8-12     5              5              +0.00%
BenchmarkTags_Pair8-12     5              5              +0.00%

benchmark                  old bytes     new bytes     delta
BenchmarkTags_Pair2-12     32            32            +0.00%
BenchmarkTags_Pair2-12     32            32            +0.00%
BenchmarkTags_Pair2-12     32            32            +0.00%
BenchmarkTags_Pair4-12     64            64            +0.00%
BenchmarkTags_Pair4-12     64            64            +0.00%
BenchmarkTags_Pair4-12     64            64            +0.00%
BenchmarkTags_Pair6-12     112           112           +0.00%
BenchmarkTags_Pair6-12     112           112           +0.00%
BenchmarkTags_Pair6-12     112           112           +0.00%
BenchmarkTags_Pair8-12     160           160           +0.00%
BenchmarkTags_Pair8-12     160           160           +0.00%
BenchmarkTags_Pair8-12     160           160           +0.00%
```